requirements.txt
streamlit
google-generativeai
python-docx
python-dotenv
beautifulsoup4
requests